8546: cleaning up component [skip ci]
